requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.6
Django==6.0.3
django-debug-toolbar==6.2.0
djangorestframework==3.16.1
pillow==12.1.1
psycopg2-binary==2.9.11
reportlab==4.4.10
sqlparse==0.5.5
